+++ main.js
@@ -19,19 +19,19 @@ const {
 const vscode = require('vscode');
 const editor = vscode.window.activeTextEditor;
 
-const firstLine = editor.document.lineAt(0);
-const lastLine = editor.document.lineAt(editor.document.lineCount - 1);
-var textRange = new vscode.Range(0,
-	firstLine.range.start.character,
-	editor.document.lineCount - 1,
-	lastLine.range.end.character);
 let HexMinifier = require('./Minifiers/hexMinifier');
 let LineRemover = require('./Minifiers/lineRemover');
 
 
-
 function activate(context) {
+	//Command MinifyAll default one.
 	const disposable = commands.registerCommand('extension.MinifyAll', () => {
+		const firstLine = editor.document.lineAt(0);
+		const lastLine = editor.document.lineAt(editor.document.lineCount - 1);
+		var textRange = new vscode.Range(0,
+			firstLine.range.start.character,
+			editor.document.lineCount - 1,
+			lastLine.range.end.character);
 		switch (window.activeTextEditor.document.languageId) {
 			case "css":
 			case "scss":
@@ -73,7 +73,6 @@ function activate(context) {
 
 				let minifierjson = new jsonMinifier(RemoverLine4Json.getLineRemoved());
 
-
 				//Get the minified code and replace it
 				let modifiedJsonText = minifierjson.getJSONMinified();
 				editor.edit(builder => {
@@ -95,14 +94,100 @@ function activate(context) {
 				break;
 
 			default:
-				window.showErrorMessage('We can not format this file type yet, use a valid one.');
+				window.showErrorMessage('⛔ We can not format this file type yet, use a valid one.');
 				break;
 		}
 
-
+		context.subscriptions.push(disposable);
 	});
 
-	context.subscriptions.push(disposable);
+	//Command MinifyAll and writes the result in other file.
+	const disposable2 = commands.registerCommand('extension.MinifyAll2OtherDoc', () => {
+		const path = require('path');
+		const FileSaver = require('fs');
+		const {
+			fileName
+		} = document;
+		const filePath = path.dirname(fileName);
+		switch (window.activeTextEditor.document.languageId) {
+			case "css":
+			case "scss":
+				const newName = path.basename(fileName).replace('.css', '-min.css');
+				const path2NewFile = path.join(filePath, newName);
+				let cssMinifier = require('./Minifiers/cssMinifier.js');
+				let cssContent = document.getText().split('\n');
+				let RemoverLine4Css = new LineRemover(cssContent);
+
+				RemoverLine4Css.removeMultipleLineComments();
+				RemoverLine4Css.removeSingleLineComments();
+				let MinifierHex4Css = new HexMinifier(RemoverLine4Css.getLineRemoved());
+
+				//Minifier processes
+				MinifierHex4Css.shortHexMain();
+				MinifierHex4Css.shortRGBMain();
+				let minifiercss = new cssMinifier(MinifierHex4Css.getHexMinified());
+
+				let modifiedCssText = minifiercss.getCssMinified();
+
+				//Get the minified code and replace it in a new file
+				FileSaver.writeFile(path2NewFile, modifiedCssText, () => {
+					window.showInformationMessage(`The minified file has been saved in: ${path2NewFile}`);
+				});
+
+				break;
+
+			case "json":
+			case "jsonc":
+				const newNamejson = path.basename(fileName).replace('.json', '-min.json');
+				const path2NewFilejson = path.join(filePath, newNamejson);
+
+				let jsonMinifier = require('./Minifiers/jsonMinifier.js');
+				let jsonContent = document.getText().split('\n');
+				let MinifierHex4Json = new HexMinifier(jsonContent);
+
+				//Minifier processes
+				MinifierHex4Json.shortHexMain();
+				MinifierHex4Json.shortRGBMain();
+
+				let RemoverLine4Json = new LineRemover(MinifierHex4Json.getHexMinified());
+
+				RemoverLine4Json.removeMultipleLineComments();
+				RemoverLine4Json.removeSingleLineComments();
+
+				let minifierjson = new jsonMinifier(RemoverLine4Json.getLineRemoved());
+				//Get the minified code and replace it in a new file
+				let modifiedJsonText = minifierjson.getJSONMinified();
+				FileSaver.writeFile(path2NewFilejson, modifiedJsonText, () => {
+					window.showInformationMessage(`The minified file has been saved in: ${path2NewFilejson}`);
+				});
+
+				break;
+
+			case "html":
+				const newNamehtml = path.basename(fileName).replace('.html', '-min.html');
+				const path2NewFilehtml = path.join(filePath, newNamehtml);
+				let htmlMinifier = require('./Minifiers/htmlMinifier.js');
+				let htmlContent = document.getText().split('\n');
+				let minifierhtml = new htmlMinifier(htmlContent);
+				minifierhtml.removeMultipleLineComments();
+
+				let modifiedHtmlText = minifierhtml.gethtmlMinified();
+
+				//Get the minified code and replace it in a new file
+				FileSaver.writeFile(path2NewFilehtml, modifiedHtmlText, () => {
+					window.showInformationMessage(`The minified file has been saved in: ${path2NewFilehtml}`);
+				});
+
+				break;
+
+			default:
+				window.showErrorMessage('⛔ We can not format this file type yet, use a valid one.');
+				break;
+		}
+
+		context.subscriptions.push(disposable2);
+	});
+
 }